Reduce amount of call stack used in hybrid_chain
Replaces `tryCatch` in `hybrid_chain` with `withCallingHandlers`. This improves
Shiny on webR on Safari < 16.4 by reducing the amount of call stack used during
deep nested calls to `hybrid_chain`.

A `delayedAssign` is used so that if the error condition handler is invoked,
the condition is handled or re-thrown in the context of `hybrid_chain`, rather
than the context of the inner function deeper in the stack.

The `finally` argument is reimplemented with an `on.exit()`.